# docs/swagger.yaml
definitions:
  dto.APIResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.AuthCustomerDTO:
    properties:
      account_type:
        example: individual
        type: string
      company_name:
        example: Tech Company Ltd
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      is_email_verified:
        example: true
        type: boolean
      is_mobile_verified:
        example: true
        type: boolean
      referrer_agency_id:
        example: 123
        type: integer
      representative_first_name:
        example: John
        type: string
      representative_last_name:
        example: Doe
        type: string
      representative_mobile:
        example: "+989123456789"
        type: string
    type: object
...
  dto.CreateCampaignResponse:
    properties:
      created_at:
        type: string
      message:
        type: string
      status:
        type: string
      uuid:
        type: string
    type: object
  dto.ForgotPasswordRequest:
    properties:
      identifier:
        example: user@example.com or +989123456789
        maxLength: 255
        minLength: 3
        type: string
    required:
    - identifier
    type: object
...
  dto.LoginRequest:
    properties:
      identifier:
        example: user@example.com or +989123456789
        maxLength: 255
        minLength: 3
        type: string
      password:
        example: SecurePass123!
        maxLength: 100
        minLength: 8
        type: string
    required:
    - identifier
    - password
    type: object
  dto.OTPResendRequest:
    properties:
      customer_id:
        type: integer
      otp_type:
        enum:
        - mobile
        - email
        type: string
    required:
    - customer_id
    - otp_type
    type: object
...
info:
  contact: {}
paths:
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Initiate password reset by sending OTP to registered mobile
      parameters:
      - description: Password reset request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP sent successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    customer_id:
                      type: integer
                    expires_in:
                      type: integer
                    masked_phone:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Forgot Password
      tags:
      - Authentication
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email/mobile and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with tokens
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    access_token:
                      type: string
                    customer:
                      $ref: '#/definitions/dto.AuthCustomerDTO'
                    expires_in:
                      type: integer
                    refresh_token:
                      type: string
                    token_type:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Authentication failed
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: User Login
      tags:
      - Authentication
  /api/v1/auth/resend-otp:
    post:
      consumes:
      - application/json
      description: Resend OTP code to user's mobile number
      parameters:
      - description: OTP resend request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.OTPResendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP resent successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    masked_otp_target:
                      type: string
                    otp_sent:
                      type: boolean
                  type: object
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Resend OTP
      tags:
      - Authentication
  /api/v1/auth/reset:
    post:
      consumes:
      - application/json
      description: Complete password reset with OTP verification
      parameters:
      - description: Password reset data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful with tokens
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    access_token:
                      type: string
                    customer:
                      $ref: '#/definitions/dto.AuthCustomerDTO'
                    expires_in:
                      type: integer
                    password_changed_at:
                      type: string
                    refresh_token:
                      type: string
                    token_type:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid request or OTP
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Reset Password
      tags:
      - Authentication
  /api/v1/auth/signup:
    post:
      consumes:
      - application/json
      description: Register a new user account with email verification
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Registration initiated successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.SignupResponse'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: User Registration
      tags:
      - Authentication
  /api/v1/auth/verify:
    post:
      consumes:
      - application/json
      description: Verify OTP code sent to user's mobile number
      parameters:
      - description: OTP verification data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.OTPVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OTP verified successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  properties:
                    access_token:
                      type: string
                    customer:
                      $ref: '#/definitions/dto.AuthCustomerDTO'
                    expires_in:
                      type: integer
                    refresh_token:
                      type: string
                    token_type:
                      type: string
                  type: object
              type: object
        "400":
          description: Invalid OTP or request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Verify OTP
      tags:
      - Authentication
  /api/v1/campaigns:
    get:
      consumes:
      - application/json
      description: Retrieve the authenticated user's campaigns with pagination, ordering,
        and filters
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Items per page (max 100)
        in: query
        name: limit
        required: true
        type: integer
      - default: newest
        description: Order by (newest|oldest)
        in: query
        name: orderby
        type: string
      - description: Filter by title (contains)
        in: query
        name: title
        type: string
      - description: Filter by status (initiated|in-progress|waiting-for-approval|approved|rejected)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ListCampaignsResponse'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: List Campaigns
      tags:
      - Campaigns
    post:
      consumes:
      - application/json
      description: Create a new campaign with the specified parameters
      parameters:
      - description: Campaign creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCampaignRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Campaign created successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CreateCampaignResponse'
              type: object
        "400":
          description: Validation error or invalid request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized - customer not found or inactive
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Create Campaign
      tags:
      - Campaigns
  /api/v1/campaigns/{uuid}:
    put:
      consumes:
      - application/json
      description: Update an existing campaign with the specified parameters
      parameters:
      - description: Campaign UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Campaign update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCampaignRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Campaign updated successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UpdateCampaignResponse'
              type: object
        "400":
          description: Validation error or invalid request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "401":
          description: Unauthorized - customer not found or inactive
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "403":
          description: Forbidden - campaign access denied or update not allowed
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: Campaign not found
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Update Campaign
      tags:
      - Campaigns
  /api/v1/campaigns/calculate-capacity:
    post:
      consumes:
      - application/json
      description: Calculate the potential reach and capacity of an campaign based
        on parameters
      parameters:
      - description: Campaign parameters for capacity calculation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CalculateCampaignCapacityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Capacity calculated successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CalculateCampaignCapacityResponse'
              type: object
        "400":
          description: Validation error or invalid request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Calculate Campaign Capacity
      tags:
      - Campaigns
  /api/v1/campaigns/calculate-cost:
    post:
      consumes:
      - application/json
      description: Calculate the total cost of an campaign based on content, target
        audience, and pricing factors
      parameters:
      - description: Campaign parameters for cost calculation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CalculateCampaignCostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Cost calculated successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CalculateCampaignCostResponse'
              type: object
        "400":
          description: Validation error or invalid request
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Calculate Campaign Cost
      tags:
      - Campaigns
...
  /api/v1/payments/callback/{invoice_number}:
    post:
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      description: Handles the callback from the payment gateway (Atipay)
      parameters:
      - description: Invoice (reservation) number
        in: path
        name: invoice_number
        required: true
        type: string
      - description: Callback data from Atipay (form or JSON)
        in: body
        name: request
        schema:
          $ref: '#/definitions/dto.AtipayRequest'
      produces:
      - text/html
      responses:
        "200":
          description: HTML payment result page
          schema:
            type: string
        "400":
          description: Invalid request or validation error
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "404":
          description: Payment request not found
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "409":
          description: Payment already processed or expired
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: Payment Callback
      tags:
      - Payments
...
swagger: "2.0"